pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django==5.2",
    "django-bootstrap5>=25.1",
    "djangorestframework",
    "djangorestframework-simplejwt",
    "django-cors-headers",
]
